+++ src/modules/cmdpal/extensions/HackerNewsExtension/LinkAction.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+using Microsoft.Windows.CommandPalette.Extensions.Helpers;
+
+namespace HackerNewsExtension;
+
+internal sealed class LinkAction : InvokableCommand
+{
+    private readonly NewsPost _post;
+
+    internal LinkAction(NewsPost post)
+    {
+        this._post = post;
+        this.Name = "Open link";
+        this.Icon = new("\uE8A7");
+    }
+
+    public override ActionResult Invoke()
+    {
+        Process.Start(new ProcessStartInfo(_post.Link) { UseShellExecute = true });
+        return ActionResult.KeepOpen();
+    }
+}
